+++ lib/src/persistence/JsonFilePersister.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:io';
+import 'dart:mirrors';
 import 'package:pip_services3_commons/pip_services3_commons.dart';
 
 import '../ILoader.dart';
@@ -24,27 +25,27 @@ import '../ISaver.dart';
 ///
 
 class JsonFilePersister<T> implements ILoader<T>, ISaver<T>, IConfigurable {
-  String path_;
+  String? path_;
 
   /// Creates a new instance of the persistence.
   ///
   /// - [path]  (optional) a path to the file where data is stored.
 
-  JsonFilePersister([String path]) {
+  JsonFilePersister([String? path]) {
     path_ = path;
   }
 
   /// Gets the file path where data is stored.
   ///
   /// Return the file path where data is stored.
-  String get path {
+  String? get path {
     return path_;
   }
 
   /// Sets the file path where data is stored.
   ///
   /// - [value]     the file path where data is stored.
-  set path(String value) {
+  set path(String? value) {
     path_ = value;
   }
 
@@ -53,7 +54,7 @@ class JsonFilePersister<T> implements ILoader<T>, ISaver<T>, IConfigurable {
   /// - [config]    configuration parameters to be set.
   @override
   void configure(ConfigParams config) {
-    path = config.getAsStringWithDefault('path', path);
+    path = config.getAsStringWithDefault('path', '');
   }
 
   /// Loads data items from external JSON file.
@@ -62,12 +63,12 @@ class JsonFilePersister<T> implements ILoader<T>, ISaver<T>, IConfigurable {
   /// Return         Future that receives loaded items
   /// Throws error.
   @override
-  Future<List<T>> load(String correlation_id) async {
+  Future<List<T>> load(String? correlation_id) async {
     if (path == null || path == '') {
-      throw ConfigException(null, 'NOpath', 'Data file path is not set');
+      throw ConfigException(null, 'NO path', 'Data file path is not set');
     }
 
-    var file = File(path);
+    var file = File(path!);
 
     if (!file.existsSync()) {
       return <T>[];
@@ -76,12 +77,22 @@ class JsonFilePersister<T> implements ILoader<T>, ISaver<T>, IConfigurable {
     try {
       var json = file.readAsStringSync();
       var list = JsonConverter.toNullableMap(json);
-      var arr = List<T>.from(ArrayConverter.listToArray(list));
+      var arr = ArrayConverter.listToArray(list);
 
-      return arr;
+      try {
+        var list = arr
+            .map((item) =>
+                reflectClass(T).newInstance(#fromJson, [item]).reflectee as T)
+            .toList();
+
+        return list;
+      } on NoSuchMethodError {
+        throw Exception(
+            'Data class must have fromJson constructor for conversions');
+      }
     } catch (ex) {
       var err = FileException(correlation_id, 'READ_FAILED',
-              'Failed to read data file: ' + path)
+              'Failed to read data file: ' + path.toString())
           .withCause(ex);
       throw err;
     }
@@ -93,14 +104,14 @@ class JsonFilePersister<T> implements ILoader<T>, ISaver<T>, IConfigurable {
   /// - [items]             list if data items to save
   /// Return         Future that error or null for success.
   @override
-  Future save(String correlation_id, List<T> items) async {
+  Future save(String? correlation_id, List<T> items) async {
     try {
       var json = JsonConverter.toJson(items);
-      File(path).writeAsStringSync(json);
+      File(path!).writeAsStringSync(json!);
       return null;
     } catch (ex) {
       var err = FileException(correlation_id, 'WRITE_FAILED',
-              'Failed to write data file: ' + path)
+              'Failed to write data file: ' + path.toString())
           .withCause(ex);
       throw err;
     }