+++ test/persistence/DummyFilePersistence_test.dart
@@ -26,5 +26,13 @@ void main() {
     test('Batch Operations', () async {
       await fixture.testBatchOperations();
     });
+
+    test('Page Sort Operations', () async {
+      await fixture.testPageSortingOperations();
+    });
+
+    test('List Sort Operations', () async {
+      await fixture.testListSortingOperations();
+    });
   });
 }